requirements.txt
# 本番環境で必要な依存関係
Flask
flask-smorest
flask-babel
Flask-Login
Flask-SQLAlchemy
Flask-Mailman
webargs
marshmallow
pillow
pillow-heif
PyMySQL
click
cryptography
pyotp
python-dotenv
qrcode
requests
celery
redis
PyJWT
Markdown
python-dateutil
MarkupSafe
gunicorn
apispec
webauthn

# 開発環境でのみ必要な依存関係
Flask-Migrate
alembic
pytest
js2py
